+++ vm/state_migration/Cargo.toml
@@ -21,6 +21,7 @@ num_cpus.workspace = true
 rayon.workspace = true
 thiserror.workspace = true
 parking_lot.workspace = true
+fvm_ipld_encoding.workspace = true
 
 fil_actor_init_v9.workspace = true
 fil_actor_init_v10.workspace = true
