+++ vm/state_migration/src/lib.rs
@@ -7,23 +7,18 @@
 use std::sync::Arc;
 
 use ahash::{HashMap, HashMapExt};
-use anyhow::anyhow;
 use cid::Cid;
-use fil_actor_init_v10::State as StateV10;
-use fil_actors_runtime_v10::runtime::EMPTY_ARR_CID;
 use forest_shim::{
     address::Address,
     state_tree::{ActorState, StateTree},
     Inner,
 };
-use forest_utils::db::BlockstoreExt;
 use fvm_ipld_blockstore::Blockstore;
 use fvm_shared::{clock::ChainEpoch, econ::TokenAmount};
 use parking_lot::Mutex;
 
-use crate::nv18::{calibnet::v10::ETH_ACCOUNT, eam::create_eam_actor};
-
-pub mod nv18;
+mod nv18;
+pub use nv18::migration::run_migration as run_nv18_migration;
 
 pub type Migrator<BS> = Arc<dyn ActorMigration<BS> + Send + Sync>;
